spliting linked list file functions
