+++ program_to_test/Makefile
@@ -6,7 +6,7 @@
 #    By: umeneses <umeneses@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/05/09 17:59:28 by umeneses          #+#    #+#              #
-#    Updated: 2024/05/23 11:48:34 by umeneses         ###   ########.fr        #
+#    Updated: 2024/05/23 14:30:41 by umeneses         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -45,8 +45,11 @@ LIBS					= $(LIBFT)
 NAME					= push_swap
 
 SRC_FILES				= push_main.c
-SRC_FILES				+= ft_utils.c
-SRC_FILES				+= ft_linkedlist.c
+SRC_FILES				+= ft_lst_addition.c
+SRC_FILES				+= ft_lst_clear.c
+SRC_FILES				+= ft_lst_deletes.c
+SRC_FILES				+= ft_lst_goto.c
+SRC_FILES				+= ft_lst_init.c
 SRC_FILES				+= ft_swap.c
 SRC_FILES				+= ft_push.c
 SRC_FILES				+= ft_rotate.c
