+++ program_to_test/src/ft_lst_goto.c
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_linkedlist.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: umeneses <umeneses@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/05/13 14:49:58 by umeneses          #+#    #+#             */
+/*   Updated: 2024/05/23 14:29:58 by umeneses         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "push_swap.h"
+
+t_stack	*ft_lst_goto_end(t_stack *list)
+{
+	if (!list)
+		return (NULL);
+	while (list && (list->next != NULL))
+		list = list->next;
+	return (list);
+}
