+++ program_to_test/src/ft_rotate.c
@@ -6,7 +6,7 @@
 /*   By: umeneses <umeneses@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/23 11:48:11 by umeneses          #+#    #+#             */
-/*   Updated: 2024/05/23 13:43:55 by umeneses         ###   ########.fr       */
+/*   Updated: 2024/05/23 14:24:43 by umeneses         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,10 +34,13 @@ void	ft_rotate_ab(t_stack **stack_a, t_stack **stack_b)
 
 void	ft_reverse_rotate(t_stack **stack)
 {
-	int	temp;
+	int		temp;
+	// t_stack	*ptr;
+
 	if (*stack == NULL || (*stack)->next == NULL)
 		return ;
 
 	temp = ft_lst_goto_end(*stack)->nbr;
 	*stack = ft_lst_addto_begin(*stack, ft_lst_init(temp));
+	// *stack = ft_lst_delat_end(*stack);
 }
